move:ProductsクラスをServicesディレクトリからCollectionsディレクトリに移動
ファーストコレクションクラスを保存するCollectionsディレクトリを作成するため。
