+++ src/app/Http/Controllers/IndexController.php
@@ -8,7 +8,7 @@ use Inertia\Inertia;
 use App\Models\Product;
 use App\Models\Brand;
 use App\Models\Category;
-use App\Services\Products;
+use App\Collections\Products;
 
 class IndexController extends Controller
 {
