+++ src/app/Http/Controllers/RankingController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\RedirectResponse;
 use App\Models\Product;
 use App\Models\Brand;
 use App\Models\Category;
-use App\Services\Products;
+use App\Collections\Products;
 use Inertia\Response;
 use Inertia\Inertia;
